// src/utils/types/BusinessResponseDto.ts
import BusinessState from '../enums/BusinessState';
import BusinessType from '../enums/BusinessType';
import CategoryClient from '../enums/CategoryClient';

type BusinessResponseDto = {
  id: string;
  title: string | null;
  numBusiness: string;
  billingDepartmentCode: string | null;
  deliveryDepartmentCode: string | null;
  billingAddressOne: string | null;
  billingAddressTwo: string | null;
  billingZipCode: string | null;
  reduction: number | null;
  billingCity: string | null;
  billingCompany: string | null;
  billingName: string | null;
  billingPhoneNumber: string | null;
  billingEmail: string | null;
  deliverAddressOne: string | null;
  deliverAddressTwo: string | null;
  deliverAddressZipCode: string | null;
  deliverAddressCity: string | null;
  deliverAddressCompany: string | null;
  deliverAddressName: string | null;
  deliverPhoneNumber: string | null;
  deliverEmail: string | null;
  enterpriseId: string;
  enterpriseName: string;
  enterpriseCategory: CategoryClient;
  numOrder: string | null;
  representativeId: string | null;
  representativeName: string | null;
  profileId: string | null;
  profileName: string | null;
  profileEmail: string | null;
  profilePhone: string | null;
  state: BusinessState | null;
  oldState: BusinessState | null;
  representativeZipCode: string | null;
  zipCode: string | null;
  deliveryMode: string | null;
  type: BusinessType | null;
  archived: boolean | null;
  installerProfileId: string | null;
  installerProfileName: string | null;
  installerProfileEmail: string | null;
  exportTva: boolean | null;
  billAndLock: boolean | null;
  billingDisabled: boolean | null;
  createdDate: Date | null;
  modifiedDate: Date | null;
  createdBy: string | null;
  modifiedBy: string | null;
};

export default BusinessResponseDto;